Move stub account info from tests to actual package
Our tests are not the only one that should be able to use it, library
users shoudl be able to import it too.
